Add `Debug` to public types from `vkutils`
This was a lot more involved than I expected.

The big change is introducing a wrapper type instead of using the
`Weak` pointers for Device and Allocator directly. This lets us
impl Debug (and potentially other traits) on it directly, and I think
this cleaned up usage of it. A lot of method chains turned into a
single `.raw()` call. I added `Self::device()` on a bunch of Vk types
too, so that getting the Arc is simpler.

The boilerplate continues to grow, but I think this adds more than it
takes away right now.

Definitely needs a sanity pass! That's why it's in its own commit.

🎉