Sprinkle basic tracing usage everywhere
This adds some basic usage of the tracing crate. It acts as a thread
safe logger and pseudo profiler at the same time.

Each run will produce a json log file to open in
    chrome://tracing
as well as events (but not spans) printed to stdout.

The tracing crate is quite extensive, but the crate we use for chrome
tracing output is somewhat lacking. If we add different tracing
outputs, we'd see more information.